requirements.txt
flask
psycopg2-binary>=2.8